Move the user preferences controller into the api namespace
