+++ .rubocop_todo.yml
@@ -15,7 +15,7 @@ Lint/AssignmentInCondition:
     - 'app/controllers/notes_controller.rb'
     - 'app/controllers/traces_controller.rb'
     - 'app/controllers/users_controller.rb'
-    - 'app/controllers/user_preferences_controller.rb'
+    - 'app/controllers/api/user_preferences_controller.rb'
     - 'app/helpers/application_helper.rb'
     - 'app/helpers/browse_helper.rb'
     - 'app/helpers/browse_tags_helper.rb'
